requirements.txt
opencv-python
imread-from-url
cap-from-youtube
onnxruntime-gpu